requirements.txt
asyncio==4.0.0
fastapi==0.121.2
loguru==0.7.3
mypy==1.18.2
numpy==2.3.5
pyaudiowpatch==0.2.12.7
pydantic-settings==2.12.0
pytest-asyncio==1.3.0
pytest==9.0.1
requests==2.32.5
ruff==0.14.5
scipy-stubs==1.16.3.0
scipy==1.16.3
sounddevice==0.5.3
types-requests==2.32.4.20250913
vosk==0.3.45